+++ mods/include/mods/feature/server.h
@@ -1,5 +1,6 @@
 FLAG(is_vanilla_compatible);
 FLAG(force_mob_spawning);
+FLAG(force_daynight_cycle);
 FLAG(generate_caves);
 FLAG(generate_tall_grass);
 FLAG(death_messages);