+++ mods/src/misc/internal.h
@@ -8,6 +8,7 @@ MCPI_INTERNAL void _init_misc_home();
 MCPI_INTERNAL void _init_misc_item_rendering();
 MCPI_INTERNAL void _init_misc_entity_rendering();
 MCPI_INTERNAL void _init_misc_chest_rendering();
+MCPI_INTERNAL void _init_misc_daynight_cycle();
 
 template <typename... Args>
 static void nop(MCPI_UNUSED Args... args) {